requirements.txt
apache-airflow==2.2.4
apache-airflow-providers-google==6.6.0

black==22.3.0
pytest==6.2.4
flake8==3.9.2
pytest-flake8==1.0.7
autopep8==1.5.7
